Move back task to Created if no time entries
